+++ dimas/src/com/publisher.rs
@@ -84,6 +84,7 @@ where
 // endregion:	--- PublisherBuilder
 
 // region:		--- Publisher
+/// Publisher
 pub struct Publisher<'a> {
 	_publisher: zenoh::publication::Publisher<'a>,
 }