+++ dimas/src/lib.rs
@@ -1,16 +1,10 @@
 // Copyright © 2023 Stephan Kunz
 #![crate_type = "lib"]
 #![crate_name = "dimas"]
-//#![warn(missing_docs)]
+//#![no_panic]
+#![warn(missing_docs)]
+#![doc = include_str!("../README.md")]
 
-//! This crate is on [crates.io](https://crates.io/crates/dimas).
-//!
-//! `DiMAS` is tested on Windows (Version 10) and Linux (Ubuntu/Debian flavours) but should also run on `MacOS`.
-//!
-//! [DiMAS](https://github.com/dimas-fw/dimas/tree/main/dimas) follows the semantic versioning principle with the enhancement,
-//! that until version 1.0.0 each new version may include breaking changes, which will be noticed in the changelog.
-
-#[doc = include_str!("../README.md")]
 #[cfg(feature = "nightly")]
 #[cfg(doctest)]
 doc_comment::doctest!("../README.md");